requirements.txt
python-dotenv==1.0.0
paramiko==3.4.0
ncclient==0.6.15
lxml==5.2.1
PyYAML==6.0.1
